+++ src/main/scala/com/ucesys/sparkscope/io/metrics/MetricReader.scala
@@ -1,9 +1,8 @@
 package com.ucesys.sparkscope.io.metrics
 
 import com.ucesys.sparkscope.data.DataTable
-import com.ucesys.sparkscope.io.MetricType
 
 trait MetricReader {
-    def readDriver(metricType: MetricType): DataTable
-    def readExecutor(metricType: MetricType, executorId: String): DataTable
+    def readDriver: DataTable
+    def readExecutor(executorId: String): DataTable
 }