+++ src/main/scala/com/ucesys/sparkscope/metrics/ExecutorMemoryStats.scala
@@ -1,6 +1,7 @@
 package com.ucesys.sparkscope.metrics
 
 import com.ucesys.sparkscope.SparkScopeAnalyzer._
+import com.ucesys.sparkscope.common.{JvmHeapMax, JvmHeapUsage, JvmHeapUsed, JvmNonHeapUsed}
 import com.ucesys.sparkscope.data.DataTable
 
 case class ExecutorMemoryStats(heapSize: Long, maxHeap: Long, maxHeapPerc: Double, avgHeap: Long, avgHeapPerc: Double, avgNonHeap: Long, maxNonHeap: Long) {
@@ -20,13 +21,13 @@ case class ExecutorMemoryStats(heapSize: Long, maxHeap: Long, maxHeapPerc: Doubl
 object ExecutorMemoryStats {
     def apply(allExecutorsMetrics: DataTable): ExecutorMemoryStats = {
         ExecutorMemoryStats(
-            heapSize = allExecutorsMetrics.select(JvmHeapMax).max.toLong / BytesInMB,
-            maxHeap = allExecutorsMetrics.select(JvmHeapUsed).max.toLong / BytesInMB,
-            maxHeapPerc = f"${allExecutorsMetrics.select(JvmHeapUsage).max}%1.5f".toDouble,
-            maxNonHeap = allExecutorsMetrics.select(JvmNonHeapUsed).max.toLong / BytesInMB,
-            avgHeap = allExecutorsMetrics.select(JvmHeapUsed).avg.toLong / BytesInMB,
-            avgHeapPerc = f"${allExecutorsMetrics.select(JvmHeapUsage).avg}%1.5f".toDouble,
-            avgNonHeap = allExecutorsMetrics.select(JvmNonHeapUsed).avg.toLong / BytesInMB
+            heapSize = allExecutorsMetrics.select(JvmHeapMax.name).max.toLong / BytesInMB,
+            maxHeap = allExecutorsMetrics.select(JvmHeapUsed.name).max.toLong / BytesInMB,
+            maxHeapPerc = f"${allExecutorsMetrics.select(JvmHeapUsage.name).max}%1.5f".toDouble,
+            maxNonHeap = allExecutorsMetrics.select(JvmNonHeapUsed.name).max.toLong / BytesInMB,
+            avgHeap = allExecutorsMetrics.select(JvmHeapUsed.name).avg.toLong / BytesInMB,
+            avgHeapPerc = f"${allExecutorsMetrics.select(JvmHeapUsage.name).avg}%1.5f".toDouble,
+            avgNonHeap = allExecutorsMetrics.select(JvmNonHeapUsed.name).avg.toLong / BytesInMB
         )
     }
 }