+++ src/test/scala/com/ucesys/sparkscope/io/metrics/LocalMetricReaderSuite.scala
@@ -19,9 +19,8 @@
 package com.ucesys.sparkscope.io.metrics
 
 import com.ucesys.sparkscope.TestHelpers._
-import com.ucesys.sparkscope.common.SparkScopeLogger
+import com.ucesys.sparkscope.common.{JvmHeapMax, JvmHeapUsage, JvmHeapUsed, JvmNonHeapUsed, SparkScopeLogger}
 import com.ucesys.sparkscope.io.file.LocalFileReader
-import com.ucesys.sparkscope.io.{JvmHeapMax, JvmHeapUsage, JvmHeapUsed, JvmNonHeapUsed}
 import org.scalamock.scalatest.MockFactory
 import org.scalatest.{FunSuite, GivenWhenThen}
 
@@ -41,15 +40,9 @@ class LocalMetricReaderSuite extends FunSuite with MockFactory with GivenWhenThe
 
         When("calling LocalMetricReader.readDriver")
         Then("LocalFileReader.read should be called with correct path")
-        (fileReaderMock.read _).expects(s"/tmp/csv-metrics/${appContext.appId}.driver.jvm.heap.used.csv").returns(jvmHeapDriverCsv)
-        (fileReaderMock.read _).expects(s"/tmp/csv-metrics/${appContext.appId}.driver.jvm.heap.usage.csv").returns(jvmHeapDriverCsv)
-        (fileReaderMock.read _).expects(s"/tmp/csv-metrics/${appContext.appId}.driver.jvm.heap.max.csv").returns(jvmHeapDriverCsv)
-        (fileReaderMock.read _).expects(s"/tmp/csv-metrics/${appContext.appId}.driver.jvm.non-heap.used.csv").returns(jvmHeapDriverCsv)
+        (fileReaderMock.read _).expects(s"/tmp/csv-metrics/${appContext.appId}/driver.csv").returns(DriverCsv)
 
-        metricsReader.readDriver(JvmHeapUsed)
-        metricsReader.readDriver(JvmHeapUsage)
-        metricsReader.readDriver(JvmHeapMax)
-        metricsReader.readDriver(JvmNonHeapUsed)
+        metricsReader.readDriver
     }
 
     test("LocalMetricReader executor metrics test") {
@@ -60,14 +53,8 @@ class LocalMetricReaderSuite extends FunSuite with MockFactory with GivenWhenThe
 
         When("calling LocalFileReader.readDriver")
         Then("LocalFileReader.read should be called with correct path")
-        (fileReaderMock.read _).expects(s"/tmp/csv-metrics/${appContext.appId}.1.jvm.heap.used.csv").returns(jvmHeapDriverCsv)
-        (fileReaderMock.read _).expects(s"/tmp/csv-metrics/${appContext.appId}.1.jvm.heap.usage.csv").returns(jvmHeapDriverCsv)
-        (fileReaderMock.read _).expects(s"/tmp/csv-metrics/${appContext.appId}.1.jvm.heap.max.csv").returns(jvmHeapDriverCsv)
-        (fileReaderMock.read _).expects(s"/tmp/csv-metrics/${appContext.appId}.1.jvm.non-heap.used.csv").returns(jvmHeapDriverCsv)
+        (fileReaderMock.read _).expects(s"/tmp/csv-metrics/${appContext.appId}/1.csv").returns(Exec1Csv)
 
-        metricsReader.readExecutor(JvmHeapUsed, "1")
-        metricsReader.readExecutor(JvmHeapUsage, "1")
-        metricsReader.readExecutor(JvmHeapMax, "1")
-        metricsReader.readExecutor(JvmNonHeapUsed, "1")
+        metricsReader.readExecutor("1")
     }
 }