+++ src/test/scala/org/apache/spark/metrics/reporter/S3CsvReporterSuite.scala
@@ -20,19 +20,33 @@ package org.apache.spark.metrics.reporter
 
 import com.codahale.metrics.MetricFilter
 import com.codahale.metrics.MetricRegistry
-import org.scalamock.scalatest.MockFactory
+import com.ucesys.sparkscope.common.SparkScopeLogger
+import com.ucesys.sparkscope.data.DataTable
+import com.ucesys.sparkscope.io.file.S3FileWriter
+import org.apache.commons.lang.SystemUtils
+import org.mockito.MockitoSugar
 import org.scalatest.{FunSuite, GivenWhenThen}
+import org.mockito.ArgumentMatchers.any
 
-import java.util.Optional
 import java.util.concurrent.ScheduledExecutorService
 import java.util.concurrent.TimeUnit;
 
-class S3CsvReporterSuite extends FunSuite with MockFactory with GivenWhenThen {
+class S3CsvReporterSuite extends FunSuite with GivenWhenThen with MockitoSugar {
 
-    def createS3CsvReporter(directory: String, region: String): S3CsvReporter = {
+    val driverMetricsStr: String =
+        """t,jvm.heap.max,jvm.heap.usage,jvm.heap.used,jvm.non-heap.used
+          |1695358645,954728448,0.29708835490780305,283638704,56840944""".stripMargin
+    val driverMetrics = DataTable.fromCsv("driver", driverMetricsStr, ",")
+
+    val exec1MetricsStr: String =
+        """t,jvm.heap.max,jvm.heap.usage,jvm.heap.used,jvm.non-heap.used,executor.cpuTime
+          |1695358697,838860800,0.21571252822875978,180952784,51120384,29815418742""".stripMargin
+
+    val exec1Metrics: DataTable = DataTable.fromCsv("1", exec1MetricsStr, ",")
+
+    def createS3CsvReporter(directory: String, writerMock: S3FileWriter = mock[S3FileWriter]): S3CsvReporter = {
         new S3CsvReporter(
             directory,
-            Optional.ofNullable(region),
             new MetricRegistry,
             null,
             ",",
@@ -41,80 +55,93 @@ class S3CsvReporterSuite extends FunSuite with MockFactory with GivenWhenThen {
             null,
             mock[MetricFilter],
             mock[ScheduledExecutorService],
-            false
-        )
+            false,
+            writerMock
+        )(mock[SparkScopeLogger])
     }
 
-    test("bucket, metricsDir, appName extraction") {
+    test("bucket, metricsDir extraction") {
         Given("s3 bucket url")
         val s3BucketUrl = "s3://my-bucket/metrics-dir"
-        And("region")
-        val region = "us-east1"
 
         When("calling S3CsvReporter constructor")
-        val s3CsvReporter = createS3CsvReporter(s3BucketUrl, region)
+        val s3CsvReporter = createS3CsvReporter(s3BucketUrl)
 
         Then("bucketName and metricsDir should be extracted")
-        assert(s3CsvReporter.bucketName.equals("my-bucket"))
-        assert(s3CsvReporter.metricsDir.equals("metrics-dir"))
-        assert(s3CsvReporter.s3 != null)
-    }
-
-    test("appName empty") {
-        Given("s3 bucket url")
-        val s3BucketUrl = "s3:///my-bucket/metrics-dir/"
-        And("region")
-        val region = "us-east1"
-
-        When("calling S3CsvReporter constructor")
-        val s3CsvReporter = createS3CsvReporter(s3BucketUrl, region)
-
-        Then("bucketName and metricsDir should be extracted")
-        assert(s3CsvReporter.bucketName.equals("my-bucket"))
-        assert(s3CsvReporter.metricsDir.equals("metrics-dir"))
-        assert(s3CsvReporter.s3 != null)
+        assert(s3CsvReporter.s3Location.bucketName.equals("my-bucket"))
+        assert(s3CsvReporter.s3Location.path.equals("metrics-dir"))
     }
 
     test("nested metricsDir") {
         Given("s3 bucket url")
         val s3BucketUrl = "s3:///my-bucket/nested-path/to/metrics-dir"
-        And("region")
-        val region = "us-east1"
 
         When("calling S3CsvReporter constructor")
-        val s3CsvReporter = createS3CsvReporter(s3BucketUrl, region)
+        val s3CsvReporter = createS3CsvReporter(s3BucketUrl)
 
         Then("bucketName and metricsDir should be extracted")
-        assert(s3CsvReporter.bucketName.equals("my-bucket"))
-        assert(s3CsvReporter.metricsDir.equals("nested-path/to/metrics-dir"))
-        assert(s3CsvReporter.s3 != null)
+        assert(s3CsvReporter.s3Location.bucketName.equals("my-bucket"))
+        assert(s3CsvReporter.s3Location.path.equals("nested-path/to/metrics-dir"))
     }
 
     test("nested metricsDir, triple slash in s3:///") {
         Given("s3 bucket url")
         val s3BucketUrl = "s3:///my-bucket/nested-path/to/metrics-dir"
         And("region")
-        val region = "us-east1"
 
         When("calling S3CsvReporter constructor")
-        val s3CsvReporter = createS3CsvReporter(s3BucketUrl, region)
+        val s3CsvReporter = createS3CsvReporter(s3BucketUrl)
 
         Then("bucketName and metricsDir should be extracted")
-        assert(s3CsvReporter.bucketName.equals("my-bucket"))
-        assert(s3CsvReporter.metricsDir.equals("nested-path/to/metrics-dir"))
-        assert(s3CsvReporter.s3 != null)
+        assert(s3CsvReporter.s3Location.bucketName.equals("my-bucket"))
+        assert(s3CsvReporter.s3Location.path.equals("nested-path/to/metrics-dir"))
     }
 
-    test("region unset") {
-        Given("s3 bucket url")
-        val s3BucketUrl = "s3:///my-bucket/nested-path/to/metrics-dir"
-        And("region")
-        val region = "us-east1"
+    test("S3CsvReporter IN_PROGRESS exists") {
+        Given("S3 IN_PROGRESS file exists")
+        val writerMock = mock[S3FileWriter]
+        doReturn(true).when(writerMock).exists(any[String])
 
-        When("calling S3CsvReporter constructor")
-        Then("IllegalArgumentException should be thrown")
-        assertThrows[IllegalArgumentException] {
-            val s3CsvReporter = createS3CsvReporter(s3BucketUrl, null)
+        And("s3 reporter")
+        val s3BucketUrl = "s3://my-bucket/metrics-dir"
+        val s3CsvReporter = createS3CsvReporter(s3BucketUrl, writerMock)
+
+        if (SystemUtils.OS_NAME == "Linux") {
+            When("s3CsvReporter.report")
+            s3CsvReporter.report("app-123-456", "driver", driverMetrics, 123)
+
+            Then("Filesystem.exists should be called")
+            verify(writerMock, times(2)).exists("s3://my-bucket/metrics-dir/.tmp/app-123-456/IN_PROGRESS")
+
+            And("And new IN_PROGRESS file should not be created")
+            verify(writerMock, times(0)).write("s3://my-bucket/metrics-dir/.tmp/app-123-456/IN_PROGRESS", "")
+
+            And("file with single row should be written")
+            verify(writerMock, times(1)).write(
+                "s3://my-bucket/metrics-dir/.tmp/app-123-456/driver/driver.123.csv",
+                driverMetrics.toCsv(",")
+            )
+        }
+    }
+
+    test("S3CsvReporter IN_PROGRESS doesn't exist ") {
+        Given("S3 IN_PROGRESS file doesn't exist")
+        val writerMock = mock[S3FileWriter]
+        doReturn(false).when(writerMock).exists(any[String])
+
+        And("s3 reporter")
+        val s3BucketUrl = "s3://my-bucket/metrics-dir"
+        val s3CsvReporter = createS3CsvReporter(s3BucketUrl, writerMock)
+
+        if (SystemUtils.OS_NAME == "Linux") {
+            When("s3CsvReporter.report")
+            s3CsvReporter.report("app-123-456", "1", exec1Metrics, 123)
+
+            Then("Filesystem.exists should be called")
+            verify(writerMock, times(2)).exists("s3://my-bucket/metrics-dir/.tmp/app-123-456/IN_PROGRESS")
+
+            And("And new IN_PROGRESS file should be created")
+            verify(writerMock, times(1)).write("s3://my-bucket/metrics-dir/.tmp/app-123-456/IN_PROGRESS", "")
         }
     }
 }