+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.1.5] - 2023-11-29 Metrics spilling optimization
+- Metrics spill refactor
+  - we are now spilling single merged csv file per instance containing all metrics instead of 5 separate files(heap used, heap used in %, heap size, non-heap used, cpuTime)
+  - local&hadoop metrics are grouped by appName+appId on storage(just like s3 metrics)
+- All sinks/reporters rewritten from Java to Scala
+
 ## [0.1.4] - 2023-11-23 UI Enhancements
 - New charts
   - chart of stages in time with number of tasks per stage